==> ch6/50.ipynb <==
import pandas as pd

FILE_PATH="./news+aggregator/newsCorpora.csv"

df = pd.read_table(FILE_PATH, names = ["ID", "TITLE", "URL", "PUBLISHER", "CATEGORY", "STORY", "HOSTNAME", "TIMESTAMP"], index_col=0)
extracted_df = df[(df["PUBLISHER"]=="Reuters")|(df["PUBLISHER"]=="Huffington Post")|(df["PUBLISHER"]=="Businessweek")|(df["PUBLISHER"]=="Contactmusic.com")|(df["PUBLISHER"]=="Daily Mail")].loc[:, ["CATEGORY", "TITLE"]] #Publisherでフィルターをかける
feature_df = extracted_df.loc[:,["TITLE", "CATEGORY"]]
feature_df.to_csv("output.txt")
shuffled_df = feature_df.sample(frac=1) #frac=1で全ての行をシャッフルする fracは選択する行の割合(つまり1だと全ての行を選択)
total_len = len(shuffled_df)
len_of_train = int(total_len * 0.8)
len_of_valid_and_test = int(total_len * 0.1)
train = shuffled_df[0:len_of_train]
valid = shuffled_df[len_of_train:len_of_train+len_of_valid_and_test]
test = shuffled_df[len_of_train+len_of_valid_and_test:]

TRAIN_FILE_PATH = "./news+aggregator/train.txt"
VALID_FILE_PATH = "./news+aggregator/valid.txt"
TEST_FILE_PATH = "./news+aggregator/test.txt"

# indexとheaderを非表示にして保存
train.to_csv(TRAIN_FILE_PATH, index=False, header=False, sep = "\t")
valid.to_csv(VALID_FILE_PATH, index=False, header=False, sep="\t")
test.to_csv(TEST_FILE_PATH, index=False, header=False, sep="\t")

# 学習，評価データのカテゴリ数を確認
train_t_num = len(train[train["CATEGORY"] == "t"])
train_e_num = len(train[train["CATEGORY"] == "e"])
train_m_num = len(train[train["CATEGORY"] == "m"])
train_b_num = len(train[train["CATEGORY"] == "b"])
test_t_num = len(test[test["CATEGORY"] == "t"])
test_e_num = len(test[test["CATEGORY"] == "e"])
test_m_num = len(test[test["CATEGORY"] == "m"])
test_b_num = len(test[test["CATEGORY"] == "b"])
data = [
    [train_t_num, test_t_num],
    [train_e_num, test_e_num],
    [train_m_num, test_m_num],
    [train_b_num, test_b_num]
]
category_df = pd.DataFrame(data=data, columns=["学習データ", "評価データ"], index=["t", "e", "m", "b"])
category_df
