==> ch6/51.ipynb <==
import pandas as pd

# 英文を単語の空白区切りに直す
def create_feature_file(file_path, output_file_path):
    df = pd.read_csv(file_path, delimiter="\t", header=None, index_col=None)
    title_words = df.iloc[:,0].replace(r'[^a-zA-Z\s]', '', regex=True)#regex = Tureで正規表現を使えるようにする
    print(title_words)
    # dfに不備があった場合への対処
    if df.shape[1] > 1:
        # 列を合体させる
        feature_df = pd.concat([title_words,df.iloc[:,1]],axis=1)
        feature_df.to_csv(output_file_path, sep="\t", index=False, header=False,)
    else:
        print(file_path)
        print("Error: Not enough columns in the DataFrame.")


TRAIN_FILE_PATH = "./news+aggregator/train.txt"
TRAIN_FEATURE_FILE_PATH = "./news+aggregator/train.feature.txt"
VALID_FILE_PATH = "./news+aggregator/valid.txt"
VALID_FEATURE_FILE_PATH = "./news+aggregator/valid.feature.txt"
TEST_FILE_PATH = "./news+aggregator/test.txt"
TEST_FEATURE_FILE_PATH = "./news+aggregator/test.feature.txt"

create_feature_file(TRAIN_FILE_PATH, TRAIN_FEATURE_FILE_PATH)
create_feature_file(VALID_FILE_PATH, VALID_FEATURE_FILE_PATH)
create_feature_file(TEST_FILE_PATH, TEST_FEATURE_FILE_PATH)
